pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crabpot"
version = "2.0.0"
description = "CrabPot — Secure OpenClaw Sandbox for WSL2"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "CrabPot Team" }]
keywords = ["openclaw", "docker", "security", "sandbox"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "flask>=3.0",
    "flask-socketio>=5.3",
    "docker>=7.0",
    "rich>=13.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.scripts]
crabpot = "crabpot.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
crabpot = ["templates/*"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
